@@ -1,10 +1,13 @@
 apply plugin: 'com.android.application'
 apply plugin: 'io.fabric'
-apply plugin: 'android-apt'
 
 repositories {
     maven { url 'https://maven.fabric.io/public' }
     maven { url "https://jitpack.io" }
+    maven {
+        url 'https://maven.google.com/'
+        name 'Google'
+    }
 }
 
 
@@ -14,9 +17,14 @@ buildscript {
         maven { url "http://dl.bintray.com/populov/maven" }
         maven { url 'https://maven.fabric.io/public' }
         jcenter()
+        maven {
+            url 'https://maven.google.com/'
+            name 'Google'
+        }
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.+'
+        classpath 'com.android.tools.build:gradle:3.4.1'
         classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
         classpath "com.android.databinding:dataBinder:1.0-rc1"
         classpath 'io.fabric.tools:gradle:1.+'
@@ -24,13 +32,12 @@ buildscript {
 }
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.0"
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "com.dev.gocar"
         minSdkVersion 21
-        targetSdkVersion 24
+        targetSdkVersion 28
         versionCode 1
         versionName "0.0.3"
     }
@@ -47,19 +54,26 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:24.2.0'
-    compile 'com.android.support:support-v4:24.2.0'
-    compile 'com.android.support:percent:24.2.0'
-    compile 'com.android.support:design:24.2.0'
-    compile 'com.squareup.picasso:picasso:2.5.2'
-    compile('com.crashlytics.sdk.android:crashlytics:2.6.2@aar') {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.12'
+
+    //Android X
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.percentlayout:percentlayout:1.0.0'
+    implementation 'androidx.coordinatorlayout:coordinatorlayout:1.1.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta4'
+
+    implementation 'com.google.android.material:material:1.1.0'
+
+    implementation 'com.squareup.picasso:picasso:2.5.2'
+    implementation('com.crashlytics.sdk.android:crashlytics:2.6.2@aar') {
         transitive = true;
     }
-    compile('com.crashlytics.sdk.android:crashlytics-ndk:1.1.4@aar') {
+    implementation('com.crashlytics.sdk.android:crashlytics-ndk:1.1.4@aar') {
         transitive = true;
     }
+    implementation 'androidx.appcompat:appcompat:1.0.0'
 }
 
 crashlytics {
