@@ -1,9 +1,10 @@
 package com.dev.gocar.ui.activities;
 
 
-import android.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.databinding.DataBindingUtil;
 
 import com.dev.gocar.R;
 import com.dev.gocar.databinding.ActivityDetailCarBinding;
