@@ -1,14 +1,13 @@
 package com.dev.gocar.ui.adapter;
 
-
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import com.dev.gocar.R;
 import com.dev.gocar.ui.model.ShowCaseModel;
-import com.dev.gocar.ui.widgets.showroom.UsedShowCase;
 import com.dev.gocar.ui.widgets.showroom.ShowCase;
+import com.dev.gocar.ui.widgets.showroom.UsedShowCase;
 
 import java.util.ArrayList;
 
@@ -19,7 +18,7 @@ public class ShowRoomAdapter extends FragmentPagerAdapter {
     public ShowRoomAdapter(FragmentManager fm, ArrayList<ShowCaseModel> showCases) {
         super(fm);
         this.showCasesModels = showCases;
-        this.showCasesModels.add(new ShowCaseModel("Carros Usados", R.drawable.showcase_02));
+        //this.showCasesModels.add(new ShowCaseModel("Carros Usados", R.drawable.showcase_02));
     }
 
     @Override
