@@ -1,8 +1,9 @@
 package com.dev.gocar.ui.pagertansformer;
 
 
-import android.support.v4.view.ViewPager;
 import android.view.View;
+
+import androidx.viewpager.widget.ViewPager;
 
 public abstract class BaseTransformer implements ViewPager.PageTransformer {
 
